rename a few file and components
